Reuse code from full-stack-tests
