+++ apps/performance-tests/package.json
@@ -14,12 +14,13 @@
     "@itwin/core-bentley": "^4.4.0",
     "@itwin/core-common": "^4.4.0",
     "@itwin/ecschema-metadata": "^4.4.0",
-    "@itwin/presentation-hierarchy-builder": "workspace:*",
     "@itwin/presentation-core-interop": "workspace:*",
+    "@itwin/presentation-hierarchy-builder": "workspace:*",
     "@itwin/presentation-models-tree": "workspace:*",
     "as-table": "^1.0.55",
-    "mocha": "^10.3.0",
     "blocked": "^1.3.0",
+    "mocha": "^10.3.0",
+    "presentation-test-utilities": "workspace:^",
     "rxjs": "^7.8.1"
   },
   "devDependencies": {
