+++ packages/full-stack-tests/package.json
@@ -68,6 +68,7 @@
     "global-jsdom": "^24.0.0",
     "mocha": "^10.3.0",
     "object-hash": "^3.0.0",
+    "presentation-test-utilities": "workspace:^",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "resize-observer-polyfill": "1.5.1",
