+++ packages/test-utilities/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "presentation-test-utilities",
+  "version": "0.0.0",
+  "private": true,
+  "scripts": {
+    "build": "tsc",
+    "clean": "rimraf lib temp",
+    "lint": "eslint \"./src/**/*.ts\""
+  },
+  "dependencies": {
+    "@itwin/core-bentley": "^4.4.0",
+    "@itwin/core-common": "^4.4.0",
+    "mocha": "^10.3.0"
+  },
+  "devDependencies": {
+    "@itwin/eslint-plugin": "4.0.0-dev.48",
+    "@types/mocha": "^10.0.6",
+    "@types/node": "^18.17.7",
+    "eslint": "^8.56.0",
+    "rimraf": "^5.0.5",
+    "typescript": "~5.0.4"
+  }
+}
